requirements.txt
pandas==1.3.3
pyodbc==4.0.32
PyPDF2==1.26.0
python-pptx==0.6.18
tesseract==0.1.3
